feat: add shared address book subpage with reusable AddressBookContent component
- Extract personal address book logic into reusable AddressBookContent component
- Add route for shared address book detail page (/address-book/shared/:guid)
- Create shared address book list page with clickable names
- Create shared address book detail page that reuses AddressBookContent
- Refactor personal address book to use the new component
- Enable navigation from shared address book list to device management subpage
